feat: Finish Default View
It has been finished the default view and, consequently,
the Category View (which is the same). Now we can see the different
categories displayed and the list of posts. It's possible to filter the
posts by category and order them by votes, name and timestamp. it has
been formatted the date of the post as well.